feat(tui): add shared Mermaid renderer
